refactor(skills): 抽取 SkillAuditSummary 为独立组件
