+++ src/renderer/pages/settings/skills/components/SkillDetailModal.tsx
@@ -18,7 +18,7 @@ import { skillHub } from '@/common/ipcBridge';
 import type { ISkillHubSkill, ISkillHubDetail } from '@/common/ipcBridge';
 import { useAppMode } from '@/renderer/hooks/useAppMode';
 import { fetchSkillDetailHttp, type SkillDetailResponse, type SkillLatestVersion } from '../index';
-import { SkillAuditSummary } from './SkillAuditReport';
+import SkillAuditSummary from './SkillAuditSummary';
 
 function parseJsonArray(jsonStr: string | null): string[] {
   if (!jsonStr) return [];
